fix(CI): split testnet tests into two files to help with flakiness
