+++ package.json
@@ -11,7 +11,7 @@
     "build": "cargo build --target wasm32-unknown-unknown --release",
     "test": "ava",
     "test:testnet": "NEAR_WORKSPACES_NETWORK=testnet yarn test",
-    "test:linkdrop": "yarn test:testnet '*/link*'",
+    "test:linkdrop": "yarn test:testnet '*/linkdrop/*'",
     "deploy:testnet": "yarn build:testnet && ts-node ./scripts/deploy.ts",
     "delpoy": "yarn build && near repl -s ./scripts/deploy.ts",
     "createKeys": "ts-node ./scripts/create-keys.ts",
